add note about maintenance status
nobody has actively worked on this branch for years, despite claims from
a number of individuals to the contrary.

until someone or a group of someones step up to spend time ensuring that
this project continues to function, I cannot recommend that anyone
tries to use it, as unmaintained code is not guaranteed to work as expected.

I've tested this today and done some backports which seem to fix startup issues,
but this should not be taken as a sign that I am returning to actively develop here
or that anything built from this branch can/should be used for an everyday desktop