e - wireless module - remove horrible memcpy with offsetof magic
while it technically wuld work, this memcpy here is fragile and the
moment some struct members changed it'd fall apart and have fields all
misaligned etc. in fact using offsetof means it might over-copy a bit
extra due to alignment of the connect_cb vs method in Connman_Service.
so manually aissgne the fields in the func instead to be far safer.

if anything this should become a sub-struct with a type that can then
be properly copied even if things change.

@fix