e start - no it doesny need dbus-launch
no. perhaps you should see the

        execvp("dbus-launch", dbus_argv);

code that auto re-launches using dbus-launch if a dbus session bus etc.
is not "found" (env vars). if your issue is that its mis-detecting the
fix the detection, but this coe went into e_start like a decade ago or
so... and it's worked every since in x11 mode and gave us a dbus
session. it SHOULD work for wayland too. don't make instructions
change and become more complex if not absoultely needed. :)