wizard - do not set scale to 1.2 forcibly. use dpi as the def prof says
the default profiel is configureed to use dpi to scale. if dpi goes up
so does wizard scaling. setitng to 1.2 forcibly is just wrong. imagine
a uhd screen thats 13" or imagine an 8k display... at least if dpi can
be read correctly things work out find. think the base dpi of 90 is
too high - then adjust that in profile... but not in wizard code.

this has been here a while and i always thought this scaling bumping
was a dpi effect. it wasnt. it was hardcoded. bad bad.

@fix.