Luncher: Add the second gadget to luncher: Luncher Grid.
Use this gadget on your desktop for a grid of icons to be used to launch applications.  The grid is simply a launcher and doesn't have all of the features that the bar has such as iconify, taskbar, styles, etc...  Please make sure to update your efl with this to get the related theme updates.