warnings - disable y2k and format non literal and float comparison
so... comapring float warnings are 98% noise and no way to shut them
up without actively changing code beavhiour. i fixed what possibly
could be issues. don't force this on for everyone as it's almost all
noise and nothing useful. turn it on in CFLAGS in your envc sometimes
when you need to double-check but dont have ti on ALL the time as
otherwise REAL warnings hide int he noise. as per e-devel discussion
this warning is at best of dubious value - i.e. very little, but it
does sometimes find something, so then let's SOMETIMEs turn it on and
not all the time.

y2k warnings are not useful for us. it's all about DISPLAy of time and
so w'ere fine.

and we use strings in config for some formats which is how it's meant
to work so these warnings are not really useful for us - at least for
now, so also turn it off.