e client mixer - stop messing with mute + volume state just for display
so add a new sink or get an update on state and e will SEt volume/mute
settings, not just passively disdplay them. this has been messing up
rage's winlist (mouse over on right) for several months now... and e
is/was wrong.

this doesnt fix all. if an app has multiple streams really this client
mixer needs to display a control per stream, not a single one - eg in
a popup. in fact volume shoud likely be done in a popup instead of
inside titlebar anyway :)

but this fixes the most annoying problem where withotu users doing
anything, the audio starts to play from streams explicitly muted by
the app...